+++ modules/compiler/src/java/flex2/compiler/mxml/builder/XMLBuilder.java
@@ -505,7 +505,7 @@ class XMLBuilder extends AbstractBuilder
             }
             else
             {
-                int num = counts.get(child.image).intValue() + 1;
+                int num = counts.get(child.image) + 1;
                 counts.put(child.image, IntegerPool.getNumber(num));
                 child.setIndex(num);
             }
